1. Реализовано отслеживание сети и сообщение пользователю об отсутствии доступа в сеть
