+++ common/src/main/java/com/example/common/data/datasource/remote/impl/TransactionRemoteDataSourceImpl.kt
@@ -4,6 +4,7 @@ import com.example.common.data.datasource.remote.TransactionRemoteDataSource
 import com.example.common.data.mapper.TransactionMapper
 import com.example.network.ResponseTemplate
 import com.example.common.domain.entity.TransactionDomain
+import com.example.network.check.NoConnectivityException
 import com.example.network.service.TransactionApiService
 import kotlinx.coroutines.delay
 import javax.inject.Inject
@@ -17,41 +18,48 @@ class TransactionRemoteDataSourceImpl @Inject constructor(
         startDate: String,
         finishDate: String
     ): ResponseTemplate<List<TransactionDomain>> {
-        var response = networkTransactionsByPeriod(accountId, startDate, finishDate)
-        repeat(3) {
-            if (response.code() == 500) {
-                delay(2000)
-                response = networkTransactionsByPeriod(accountId, startDate, finishDate)
-            } else return@repeat
-        }
-        return when (response.code()) {
-            200, 201, 204 -> ResponseTemplate(
-                typeResponse = ResponseTemplate.TypeResponse.SUCCESS,
-                body = response.body()?.map { transactionMapper.toTransactionDomain(it) }
-            )
+        try {
+            var response = networkTransactionsByPeriod(accountId, startDate, finishDate)
+            repeat(3) {
+                if (response.code() == 500) {
+                    delay(2000)
+                    response = networkTransactionsByPeriod(accountId, startDate, finishDate)
+                } else return@repeat
+            }
+            return when (response.code()) {
+                200, 201, 204 -> ResponseTemplate(
+                    typeResponse = ResponseTemplate.TypeResponse.SUCCESS,
+                    body = response.body()?.map { transactionMapper.toTransactionDomain(it) }
+                )
 
-            400 -> ResponseTemplate(
-                typeResponse = ResponseTemplate.TypeResponse.ERROR_CLIENT,
-                body = null
-            )
+                400 -> ResponseTemplate(
+                    typeResponse = ResponseTemplate.TypeResponse.ERROR_CLIENT,
+                    body = null
+                )
 
-            401 -> ResponseTemplate(
-                typeResponse = ResponseTemplate.TypeResponse.UNAUTHORIZED,
-                body = null
-            )
+                401 -> ResponseTemplate(
+                    typeResponse = ResponseTemplate.TypeResponse.UNAUTHORIZED,
+                    body = null
+                )
 
-            404 -> ResponseTemplate(
-                typeResponse = ResponseTemplate.TypeResponse.NOT_FOUND,
-                body = null
-            )
+                404 -> ResponseTemplate(
+                    typeResponse = ResponseTemplate.TypeResponse.NOT_FOUND,
+                    body = null
+                )
 
-            500 -> ResponseTemplate(
-                typeResponse = ResponseTemplate.TypeResponse.ERROR_SERVER,
-                body = null
-            )
+                500 -> ResponseTemplate(
+                    typeResponse = ResponseTemplate.TypeResponse.ERROR_SERVER,
+                    body = null
+                )
 
-            else -> ResponseTemplate(
-                typeResponse = ResponseTemplate.TypeResponse.ALL_BAD,
+                else -> ResponseTemplate(
+                    typeResponse = ResponseTemplate.TypeResponse.ALL_BAD,
+                    body = null
+                )
+            }
+        } catch (e: NoConnectivityException) {
+            return ResponseTemplate(
+                typeResponse = ResponseTemplate.TypeResponse.NETWORK_PROBLEM,
                 body = null
             )
         }
