+++ common/src/main/java/com/example/common/presentation/base_visible/ErrorVisible.kt
@@ -21,11 +21,13 @@ fun ErrorVisible(
             Text(message)
         } else {
             when (type) {
+                ResponseTemplate.TypeResponse.SUCCESS -> Text("Данные есть, но что-то пошло не так")
                 ResponseTemplate.TypeResponse.UNAUTHORIZED -> Text("Ошибка авторизации")
                 ResponseTemplate.TypeResponse.NOT_FOUND -> Text("Ошибка в данных")
                 ResponseTemplate.TypeResponse.ERROR_SERVER -> Text("Внутренняя ошибка сервера")
                 ResponseTemplate.TypeResponse.ERROR_CLIENT -> Text("Внутренняя ошибка приложения")
-                else -> Text("Неизвесная ошибка")
+                ResponseTemplate.TypeResponse.NETWORK_PROBLEM -> Text("Проблемы с сетью.\nПроверьте подключение к интернету")
+                ResponseTemplate.TypeResponse.ALL_BAD -> Text("Неизвесная ошибка")
             }
         }
     }
