+++ features/category/src/main/java/com/example/category/data/datasource/remote/impl/CategoryRemoteDataSourceImpl.kt
@@ -4,6 +4,7 @@ import com.example.common.data.mapper.CategoryMapper
 import com.example.network.ResponseTemplate
 import com.example.category.data.datasource.remote.CategoryRemoteDataSource
 import com.example.common.domain.entity.CategoryDomain
+import com.example.network.check.NoConnectivityException
 import com.example.network.service.CategoryApiService
 import kotlinx.coroutines.delay
 import javax.inject.Inject
@@ -13,41 +14,48 @@ class CategoryRemoteDataSourceImpl @Inject constructor(
     private val categoryApiService: CategoryApiService
 ) : CategoryRemoteDataSource {
     override suspend fun getCategories(): ResponseTemplate<List<CategoryDomain>> {
-        var response = networkCategories()
-        repeat(3) {
-            if (response.code() == 500) {
-                delay(2000)
-                response = networkCategories()
-            } else return@repeat
-        }
-        return when (response.code()) {
-            200, 201, 204 -> ResponseTemplate(
-                typeResponse = ResponseTemplate.TypeResponse.SUCCESS,
-                body = response.body()?.map { categoryMapper.toCategoryDomain(it) }
-            )
+        try {
+            var response = networkCategories()
+            repeat(3) {
+                if (response.code() == 500) {
+                    delay(2000)
+                    response = networkCategories()
+                } else return@repeat
+            }
+            return when (response.code()) {
+                200, 201, 204 -> ResponseTemplate(
+                    typeResponse = ResponseTemplate.TypeResponse.SUCCESS,
+                    body = response.body()?.map { categoryMapper.toCategoryDomain(it) }
+                )
 
-            400 -> ResponseTemplate(
-                typeResponse = ResponseTemplate.TypeResponse.ERROR_CLIENT,
-                body = null
-            )
+                400 -> ResponseTemplate(
+                    typeResponse = ResponseTemplate.TypeResponse.ERROR_CLIENT,
+                    body = null
+                )
 
-            401 -> ResponseTemplate(
-                typeResponse = ResponseTemplate.TypeResponse.UNAUTHORIZED,
-                body = null
-            )
+                401 -> ResponseTemplate(
+                    typeResponse = ResponseTemplate.TypeResponse.UNAUTHORIZED,
+                    body = null
+                )
 
-            404 -> ResponseTemplate(
-                typeResponse = ResponseTemplate.TypeResponse.NOT_FOUND,
-                body = null
-            )
+                404 -> ResponseTemplate(
+                    typeResponse = ResponseTemplate.TypeResponse.NOT_FOUND,
+                    body = null
+                )
 
-            500 -> ResponseTemplate(
-                typeResponse = ResponseTemplate.TypeResponse.ERROR_SERVER,
-                body = null
-            )
+                500 -> ResponseTemplate(
+                    typeResponse = ResponseTemplate.TypeResponse.ERROR_SERVER,
+                    body = null
+                )
 
-            else -> ResponseTemplate(
-                typeResponse = ResponseTemplate.TypeResponse.ALL_BAD,
+                else -> ResponseTemplate(
+                    typeResponse = ResponseTemplate.TypeResponse.ALL_BAD,
+                    body = null
+                )
+            }
+        } catch (e: NoConnectivityException) {
+            return ResponseTemplate(
+                typeResponse = ResponseTemplate.TypeResponse.NETWORK_PROBLEM,
                 body = null
             )
         }
