--- network/src/main/java/com/example/network/NetworkConnectionInterceptor.kt
@@ -1,33 +1,0 @@
-package com.example.network
-
-import android.content.Context
-import android.net.ConnectivityManager
-import android.net.NetworkCapabilities
-import okhttp3.Interceptor
-import okhttp3.Response
-import java.io.IOException
-
-class NetworkConnectionInterceptor(
-    private val context: Context
-): Interceptor {
-    override fun intercept(chain: Interceptor.Chain): Response {
-        if (isNetworkAvailable(context).not()) {
-            throw NoConnectivityException()
-        }
-        return chain.proceed(chain.request())
-    }
-}
-
-class NoConnectivityException: IOException("Нет подключения к интернету")
-
-fun isNetworkAvailable(context: Context): Boolean {
-    val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-    val activeNetwork = connectivityManager.activeNetwork ?: return false
-    val networkCapabilities = connectivityManager.getNetworkCapabilities(activeNetwork) ?: return false
-    return when {
-        networkCapabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> true
-        networkCapabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> true
-        networkCapabilities.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET) -> true
-        else -> false
-    }
-}
