+++ network/src/main/java/com/example/network/check/NetworkConnectionInterceptor.kt
@@ -1,0 +1,16 @@
+package com.example.network.check
+
+import android.content.Context
+import okhttp3.Interceptor
+import okhttp3.Response
+
+class NetworkConnectionInterceptor(
+    private val context: Context
+) : Interceptor {
+    override fun intercept(chain: Interceptor.Chain): Response {
+        if (isNetworkAvailable(context).not()) {
+            throw NoConnectivityException()
+        }
+        return chain.proceed(chain.request())
+    }
+}
