+++ network/src/main/java/com/example/network/check/NoConnectivityException.kt
@@ -1,0 +1,5 @@
+package com.example.network.check
+
+import java.io.IOException
+
+class NoConnectivityException : IOException("Нет подключения к интернету")
